backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
autopep8==2.3.2
python-dotenv==1.1.1
psycopg[binary,pool]
inflect==7.5.0
sqlparse==0.5.3
flask-shell-ipython==0.5.3
validate_email==1.3
Flask-JWT-Extended==4.7.1
boto3==1.40.8
cryptography==45.0.6
pytz==2025.2
flask-cors==6.0.1
Faker==37.6.0
html2text==2025.4.15
